# api/openapi.yaml
openapi: 3.0.3
info:
  title: Simplus API
  version: 0.1.0
  description: Locale-neutral control-plane contract for Simplus.
servers:
  - url: /
paths:
...
  /api/v1/auth/session:
    get:
      operationId: getAuthSession
      summary: Read the current administrator browser session
      tags: [auth]
      responses:
        "200":
          description: Current authenticated administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Session state could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/v1/auth/password:
    put:
      operationId: changeAdministratorPassword
      summary: Replace the sole administrator password and revoke all sessions
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeAdministratorPasswordRequest"
      responses:
        "204": { description: Password replaced; every administrator session was revoked }
        "400":
          description: Password request is invalid
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "401":
          description: Session or current password is invalid
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: CSRF validation failed
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "500":
          description: Password could not be replaced
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
...
  schemas:
...
    ChangeAdministratorPasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword, newPasswordConfirmation]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 1024 }
        newPassword: { type: string, minLength: 12, maxLength: 256 }
        newPasswordConfirmation: { type: string, minLength: 12, maxLength: 256 }
...
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required: [username, locale, expiresAt]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        locale:
          type: string
          enum: [zh-CN, en-US]
        expiresAt:
          type: string
          format: date-time
    ApiError:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
        retryable:
          type: boolean
        reference:
          type: string